Clean up tests/lib.rs
