+++ tests/test.rs
@@ -1,5 +1,35 @@
 use super::*;
 
+#[derive(Debug)]
+enum Expected {
+  String(String),
+  Regex(Regex),
+  Ignore,
+}
+
+impl Expected {
+  fn regex(pattern: &str) -> Self {
+    Self::Regex(Regex::new(&format!("^(?s){}$", pattern)).unwrap())
+  }
+
+  fn assert_match(&self, output: &str) {
+    match self {
+      Self::String(string) => assert_eq!(output, string),
+      Self::Regex(regex) => assert!(
+        regex.is_match(output),
+        "output did not match regex: {:?}",
+        output
+      ),
+      Self::Ignore => {}
+    }
+  }
+}
+
+pub(crate) struct Output {
+  pub(crate) stdout: String,
+  pub(crate) state: State,
+}
+
 pub(crate) struct Test {
   args: Vec<String>,
   expected_status: i32,
