+++ tests/transaction_options.rs
@@ -1,0 +1,5 @@
+pub(crate) struct TransactionOptions<'a> {
+  pub(crate) slots: &'a [(usize, usize, usize)],
+  pub(crate) output_count: usize,
+  pub(crate) fee: u64,
+}
